change logic to use pushy command uniformly
