+++ src/modules/user-module.ts
@@ -1,71 +1,12 @@
 import { getSession, loadSession } from '../api';
-import type { CLIModule, CommandContext, CommandResult } from '../types';
+import type { CLIModule, CommandContext } from '../types';
 import { userCommands } from '../user';
 
 export const userModule: CLIModule = {
   name: 'user',
   version: '1.0.0',
 
-  commands: [
-    {
-      name: 'login',
-      description: 'Login to the service',
-      handler: async (context: CommandContext): Promise<CommandResult> => {
-        try {
-          console.log('Logging in user');
-          await userCommands.login(context);
-          return {
-            success: true,
-            data: { message: 'Login successful' },
-          };
-        } catch (error) {
-          return {
-            success: false,
-            error: error instanceof Error ? error.message : 'Login failed',
-          };
-        }
-      },
-    },
-    {
-      name: 'logout',
-      description: 'Logout from the service',
-      handler: async (context: CommandContext): Promise<CommandResult> => {
-        try {
-          console.log('Logging out user');
-          await userCommands.logout(context);
-          return {
-            success: true,
-            data: { message: 'Logout successful' },
-          };
-        } catch (error) {
-          return {
-            success: false,
-            error: error instanceof Error ? error.message : 'Logout failed',
-          };
-        }
-      },
-    },
-    {
-      name: 'me',
-      description: 'Show current user information',
-      handler: async (context: CommandContext): Promise<CommandResult> => {
-        try {
-          console.log('Getting user information');
-          await userCommands.me();
-          return {
-            success: true,
-            data: { message: 'User information retrieved successfully' },
-          };
-        } catch (error) {
-          return {
-            success: false,
-            error:
-              error instanceof Error ? error.message : 'Get user info failed',
-          };
-        }
-      },
-    },
-  ],
+  commands: [],
 
   workflows: [
     {
